feat(website): turn the mobile drawer into the full-width navigation panel
- move backdrop-filter off .navbar: it made the fixed drawer clip to the navbar height
- share header, section switcher, search and footer between the desktop sidebar and the drawer
- replace the hover-only locale dropdown with a click-toggled one that works on touch
