+++ Website/src/components/NavigationPanel/styles.module.css
@@ -1,10 +1,16 @@
 .panel { display: flex; flex-direction: column; width: var(--doc-sidebar-width); height: 100%; min-height: 0; }
+/* The mobile drawer fills the viewport; its height is the drawer's, so the list scrolls between header and footer. */
+.panelMobile { width: 100%; height: 100dvh; }
 
 /* Everything shares the menu's edges: boxes start at 0.5rem (menu padding), text and icons at 1.25rem (link padding). */
 .header { flex: none; display: flex; flex-direction: column; gap: 0.75rem; margin: 0 0.5rem; padding: 0.75rem 0 0.85rem; border-bottom: 1px solid var(--venom-line); }
 .tools { display: flex; align-items: stretch; gap: 0.5rem; }
 .tools > :first-child { flex: 1 1 auto; min-width: 0; }
 .tools > button { flex: none; width: 2.5rem; justify-content: center; border-radius: 4px; }
+.brandRow { display: flex; align-items: center; justify-content: space-between; gap: 0.5rem; }
+.close { display: inline-flex; align-items: center; justify-content: center; width: 2.5rem; height: 2.5rem; margin-right: -0.5rem; color: var(--venom-fg); border-radius: 4px; }
+.close:hover { color: var(--venom-accent); background: var(--venom-glow-soft); }
+.close svg { width: 22px; height: 22px; }
 .brandLink { display: flex; align-items: center; gap: 0.6rem; min-width: 0; padding: 0.25rem 0; color: var(--venom-fg-strong); }
 .brandLink:hover { color: var(--venom-fg-strong); text-decoration: none; }
 .brandLogo { flex: none; display: flex; height: 1.75rem; }
@@ -17,6 +23,7 @@
 .switcherButton > svg { flex: none; color: var(--venom-muted); transition: transform var(--ifm-transition-fast); }
 .switcherButton[aria-expanded='true'] > svg { transform: rotate(180deg); }
 .switcherMenu { position: absolute; z-index: 20; inset: calc(100% + 0.35rem) 0 auto; margin: 0; padding: 0.35rem; list-style: none; background: var(--venom-panel); border: 1px solid var(--venom-line-strong); border-radius: 6px; box-shadow: 0 12px 40px #0005; }
+.switcherMenuUp { inset: auto 0 calc(100% + 0.35rem) auto; min-width: 10rem; }
 .switcherItem { display: block; padding: 0.45rem 0.75rem; border-radius: 4px; color: var(--venom-fg); font-size: 0.875rem; line-height: 1.5; }
 .switcherItem:hover { color: var(--venom-fg-strong); background: var(--venom-glow-soft); text-decoration: none; }
 .switcherItemActive, .switcherItemActive:hover { color: var(--venom-accent); font-weight: 600; }
@@ -27,12 +34,8 @@
 .list :global(.menu) { flex: 1 1 auto; min-height: 0; overflow-y: auto; }
 
 .footer { flex: none; display: flex; align-items: center; justify-content: space-between; gap: 0.25rem; margin: 0 0.5rem; padding: 0.5rem 0.375rem; border-top: 1px solid var(--venom-line); }
-.icon, .footer :global(.clean-btn), .locale { display: inline-flex; align-items: center; justify-content: center; width: 2rem; height: 2rem; padding: 0; color: var(--venom-fg); border-radius: 4px; }
-.icon:hover, .footer :global(.clean-btn:hover), .locale:hover { color: var(--venom-accent); background: var(--venom-glow-soft); text-decoration: none; }
-.footer :global(.navbar__item.dropdown) { position: relative; display: block; margin: 0; padding: 0; }
-.locale { font-size: 0; }
-.locale::after { display: none; }
-.locale svg { width: 20px; height: 20px; margin: 0; }
-.footer :global(.dropdown__menu) { top: auto; bottom: calc(100% + 0.35rem); left: 0; right: auto; min-width: 10rem; background: var(--venom-panel); border: 1px solid var(--venom-line-strong); border-radius: 6px; box-shadow: 0 12px 40px #0005; }
+.icon, .footer :global(.clean-btn), .locale { display: inline-flex; align-items: center; justify-content: center; width: 2rem; height: 2rem; padding: 0; color: var(--venom-fg); background: transparent; border: 0; border-radius: 4px; cursor: pointer; }
+.icon:hover, .footer :global(.clean-btn:hover), .locale:hover, .locale[aria-expanded='true'] { color: var(--venom-accent); background: var(--venom-glow-soft); text-decoration: none; }
+.locale svg { width: 20px; height: 20px; }
 .theme :global(button) { color: var(--venom-fg); }
 .theme :global(svg) { width: 20px; height: 20px; }
